replace the formik-auto-submit component with the new one
