+++ src/components/form/index.ts
@@ -8,5 +8,4 @@ export * from './Select'
 export * from './SelectMonth'
 export * from './SelectYear'
 export * from './TextArea'
-export * from './AutoSubmit'
 export * from './useCustomInputCss'
